rename charmbits *charm to *relation
